+++ src/main/java/com/example/onlinebookstore/service/impl/CategoryServiceImpl.java
@@ -48,8 +48,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto updateById(CreateCategoryRequestDto requestDto, Long id) {
-        categoryRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Can't find category by id: " + id));
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find category by id: " + id);
+        }
         Category model = categoryMapper.toModel(requestDto);
         model.setId(id);
         return categoryMapper.toDto(categoryRepository.save(model));
@@ -57,8 +58,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
-        categoryRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Can't find category by id: " + id));
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find category by id: " + id);
+        }
         categoryRepository.deleteById(id);
     }
